Fixan integer overflow bug that goes back at least as far as jai-imageio.  for  the fix in the java version see: https://github.com/faceless2/jpeg2000/commit/b20342956ac5d05bb97522ddeb41e9bed5abfa95
The code overflows an int in bp (bitplanes) is 30.  They have a complicated fix.  I just do the shifting in a long (which is essentially a free conversion on a 64 bit processor) and then cast it back to int when  the shifting is done.  (Also essentially free on most 64 bit processors.)

+++ CoreJ2K/j2k/entropy/decoder/StdEntropyDecoder.cs
@@ -910,7 +910,7 @@ namespace CoreJ2K.j2k.entropy.decoder
 			sscanw = cblk.w + 2;
 			jstep = sscanw * StdEntropyCoderOptions.STRIPE_HEIGHT / 2 - cblk.w;
 			kstep = dscanw * StdEntropyCoderOptions.STRIPE_HEIGHT - cblk.w;
-			setmask = (3 << bp) >> 1;
+			setmask = (int)(((long)3<<bp)>>1);
 			data = (int[]) cblk.Data;
 			nstripes = (cblk.h + StdEntropyCoderOptions.STRIPE_HEIGHT - 1) / StdEntropyCoderOptions.STRIPE_HEIGHT;
 			causal = (options & StdEntropyCoderOptions.OPT_VERT_STR_CAUSAL) != 0;
@@ -1224,7 +1224,7 @@ namespace CoreJ2K.j2k.entropy.decoder
 			sscanw = cblk.w + 2;
 			jstep = sscanw * StdEntropyCoderOptions.STRIPE_HEIGHT / 2 - cblk.w;
 			kstep = dscanw * StdEntropyCoderOptions.STRIPE_HEIGHT - cblk.w;
-			setmask = (3 << bp) >> 1;
+			setmask = (int)(((long)3<<bp)>>1);
 			data = (int[]) cblk.Data;
 			nstripes = (cblk.h + StdEntropyCoderOptions.STRIPE_HEIGHT - 1) / StdEntropyCoderOptions.STRIPE_HEIGHT;
 			causal = (options & StdEntropyCoderOptions.OPT_VERT_STR_CAUSAL) != 0;
@@ -1865,7 +1865,7 @@ namespace CoreJ2K.j2k.entropy.decoder
 			sscanw = cblk.w + 2;
 			jstep = sscanw * StdEntropyCoderOptions.STRIPE_HEIGHT / 2 - cblk.w;
 			kstep = dscanw * StdEntropyCoderOptions.STRIPE_HEIGHT - cblk.w;
-			setmask = (3 << bp) >> 1;
+			setmask = (int)(((long)3<<bp)>>1);
 			data = (int[]) cblk.Data;
 			nstripes = (cblk.h + StdEntropyCoderOptions.STRIPE_HEIGHT - 1) / StdEntropyCoderOptions.STRIPE_HEIGHT;
 			causal = (options & StdEntropyCoderOptions.OPT_VERT_STR_CAUSAL) != 0;
